extracted createEmptyGrid, renamed createInitialGrid test utility to create grid, fixed ts errors in tests by renaming js to ts extensions in tests files
