+++ src/components/utils/checkUtils/__tests__/checkSecondary.test.ts
@@ -1,15 +1,15 @@
 import { expect, test, describe } from 'vitest'
-import createInitialGrid from '../utils'
-import checkPrimary from '../checkPrimaryDiagonal'
 import { X } from '@/constants'
+import createGrid from '../createGrid'
+import checkSecondary from '../checkSecondaryDiagonal'
 
 const lookingFor = X
 const startRowIdx = 0
 const startColIdx = 0
 const target = 3
 
-describe('checkPrimary', () => {
-  test('should check a primary with false result', () => {
+describe('checkSecondary', () => {
+  test('should check a secondary with false result', () => {
     const emptyGridOptions = {
       N: 4,
       winIndex: 0,
@@ -18,8 +18,8 @@ describe('checkPrimary', () => {
       mode: '',
       hasWinningMarks: false
     }
-    const grid = createInitialGrid(emptyGridOptions)
-    const result = checkPrimary({ grid, startRowIdx, startColIdx, lookingFor, target })
+    const grid = createGrid(emptyGridOptions)
+    const result = checkSecondary({ grid, startRowIdx, startColIdx, lookingFor, target })
 
     expect(result).toEqual({
       winner: '',
@@ -27,18 +27,18 @@ describe('checkPrimary', () => {
     })
   })
 
-  test('should check primary with false as winning result', () => {
+  test('should check secondary with false as winning result', () => {
     const gridOptions = {
       N: 3,
       winIndex: 0,
       start: 0,
       end: 1,
-      mode: 'primary',
+      mode: 'secondary',
       hasWinningMarks: false
     }
-    const grid = createInitialGrid(gridOptions)
+    const grid = createGrid(gridOptions)
 
-    const result = checkPrimary({
+    const result = checkSecondary({
       grid,
       startRowIdx,
       startColIdx,
@@ -53,21 +53,21 @@ describe('checkPrimary', () => {
     })
   })
 
-  test('should check primary with true as winning result', () => {
+  test('should check secondary with true as winning result', () => {
     const gridOptions = {
       N: 3,
       winIndex: 0,
       start: 0,
       end: 2,
-      mode: 'primary',
+      mode: 'secondary',
       hasWinningMarks: false
     }
-    const grid = createInitialGrid(gridOptions)
+    const grid = createGrid(gridOptions)
 
-    const result = checkPrimary({
+    const result = checkSecondary({
       grid,
       startRowIdx,
-      startColIdx,
+      startColIdx: gridOptions.N - 1,
       lookingFor,
       target,
       shouldMarkWinningCells: true
@@ -79,24 +79,24 @@ describe('checkPrimary', () => {
     })
 
     gridOptions.hasWinningMarks = true
-    expect(grid).toEqual(createInitialGrid(gridOptions))
+    expect(grid).toEqual(createGrid(gridOptions))
   })
 
-  test('should check primary with true as winning result in a subrange', () => {
+  test('should check secondary with true as winning result in a subrange', () => {
     const gridOptions = {
       N: 4,
       winIndex: 0,
       start: 1,
       end: 3,
-      mode: 'primary',
+      mode: 'secondary',
       hasWinningMarks: false
     }
-    const grid = createInitialGrid(gridOptions)
+    const grid = createGrid(gridOptions)
 
-    const result = checkPrimary({
+    const result = checkSecondary({
       grid,
-      startRowIdx: gridOptions.winIndex,
-      startColIdx,
+      startRowIdx: 1,
+      startColIdx: 2,
       lookingFor,
       target,
       shouldMarkWinningCells: true
@@ -108,6 +108,6 @@ describe('checkPrimary', () => {
     })
 
     gridOptions.hasWinningMarks = true
-    expect(grid).toEqual(createInitialGrid(gridOptions))
+    expect(grid).toEqual(createGrid(gridOptions))
   })
 })
