+++ src/components/utils/checkUtils/createGrid.ts
@@ -1,7 +1,7 @@
 import { X, W } from '@/constants'
 import type { gridT, rowT } from '@/components/types'
 
-interface createInitialGridI {
+interface createGridI {
   N: number
   winIndex: number
   start: number
@@ -15,7 +15,7 @@ interface createInitialGridI {
  * It will create a 2D array with marks and symbols in various subranges that represents
  * user's selection. This utility function is used by testing to check winning  results
  * */
-const createInitialGrid = ({
+const createGrid = ({
   N,
   winIndex,
   start,
@@ -24,7 +24,7 @@ const createInitialGrid = ({
   hasWinningMarks,
   lookingFor = X,
   emptyString = ''
-}: createInitialGridI): gridT => {
+}: createGridI): gridT => {
   const grid: gridT = []
 
   for (let i = 0; i < N; i++) {
@@ -59,4 +59,4 @@ const createInitialGrid = ({
 
   return grid
 }
-export default createInitialGrid
+export default createGrid
